backend/requirements.txt
Django==3.2.16
djangorestframework==3.13.1
django-cors-headers==3.11.0
Pillow==9.3.0
